Mon, Jan 17, 2022, 6:47 PM +00:00

+++ filter-less/helpers.c
@@ -131,9 +131,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
                     }
                 }
             }
-            int avgBlue = round(totalBlue / counter);
-            int avgGreen = round(totalGreen / counter);
-            int avgRed = round(totalRed / counter);
+            int avgBlue = round((float)totalBlue / counter);
+            int avgGreen = round((float)totalGreen / counter);
+            int avgRed = round((float)totalRed / counter);
 
             if (avgBlue > 255)
             {
